tests/wpt/tests/css/css-rhythm/parsing/block-step-invalid.html

<!DOCTYPE html>
<html>
<head>
<title>CSS Rhythm: block-step invalid values</title>
<link rel="author" title="Tim Nguyen" href="https://github.com/nt1m">
<link rel="help" href="https://drafts.csswg.org/css-rhythm/#block-step">
<meta name="assert" content="Invalid values for block-step should not be parsed">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
    test_invalid_value("block-step", "auto auto");
    test_invalid_value("block-step", "start end");
    test_invalid_value("block-step", "center start");
    test_invalid_value("block-step", "none none");
    test_invalid_value("block-step", "300px none");
    test_invalid_value("block-step", "300px start none");
    test_invalid_value("block-step", "300px start border-box padding-box");
</script>
</body>
</html>
